requirements.txt
Django==2.1.7
django-markdownx==2.0.28
django-rest-swagger==2.2.0
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docker-compose==1.23.2
flake8==3.7.7
gunicorn==19.9.0
ldap3==2.5.2
Markdown==3.1.1
MarkupSafe>=1.1.1
Pillow==6.0.0
psycopg2-binary
pyflakes==2.1.1
easydict==1.9
django-auditlog==0.4.7
gevent==21.1.2
django-markdownify==0.9.0
xhtml2pdf==0.2.5
reportlab==3.6.5
